requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
torch>=2.0.0
transformers>=4.35.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
pillow>=10.0.0
matplotlib>=3.7.0
numpy>=1.24.0 
python-multipart
google-generativeai>=0.3.0
python-dotenv>=1.0.0
google-generativea